+++ erdst/realcal.F90
@@ -48,8 +48,8 @@ module realcal
   integer :: block_size(3)
   
   ! "straight" coordinate system
-  real, allocatable :: sitepos_normal(:, :)
-  real :: cell_normal(3, 3), invcell_normal(3), cell_len_normal(3)
+  real(kind=8), allocatable :: sitepos_normal(:, :)
+  real(kind=8) :: cell_normal(3, 3), invcell_normal(3), cell_len_normal(3)
   real :: invcell(3, 3)
   logical :: is_cuboid
   real, parameter :: check_rotate = 1e-8, cuboid_thres = 1e-8
